requirements.txt
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
google-cloud-aiplatform>=1.38.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
dropbox>=11.36.0
